moving information about the router
