+++ src/main.js
@@ -6,6 +6,10 @@ import { createVuetify } from 'vuetify'
 import * as components from 'vuetify/components'
 import * as directives from 'vuetify/directives'
 
+// Router
+import { createMemoryHistory, createRouter } from 'vue-router'
+import ListPerson from '../components/ListPerson.vue'
+
 // Components
 import App from './App.vue'
 
@@ -14,4 +18,14 @@ const vuetify = createVuetify({
   directives,
 })
 
-createApp(App).use(vuetify).mount('#app')
+const routes = [
+  { path: '/', component: App },
+  { path: '/persons', component: ListPerson },
+]
+
+const router = createRouter({
+  history: createMemoryHistory(),
+  routes,
+})
+
+createApp(App).use(router).use(vuetify).mount('#app')
